marks in Terminal, small changes

+++ src/06_d-Heap/dHeap_lib/Includes/Edge.cpp
@@ -6,7 +6,7 @@ Edge::Edge()
 	weight = start = end = 0;
 };
 
-Edge::Edge(int _w, int _s, int _e)
+Edge::Edge(float _w, int _s, int _e)
 {
 	if (_w < 0 ||_s < 0 ||_e < 0)
 		throw Exception("Incorrect edge's data!");
